% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{average_age_survivors}
\alias{average_age_survivors}
\title{: Calculate Average Age of Survivors}
\usage{
average_age_survivors(age, survival)
}
\arguments{
\item{age:}{Vector of passenger ages}

\item{survival:}{Vector of survival status}
}
\value{
None: Prints average age of survivors
}
\description{
: Computes and prints the average age of passengers who survived
}
\examples{
# Calculate average age of survivors
average_age_survivors(age, survival)
}
