% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{average_health_by_site}
\alias{average_health_by_site}
\title{: Average Health Score by Crash Site}
\usage{
average_health_by_site(health_score, crash_site)
}
\arguments{
\item{health_score:}{Vector of health scores}

\item{crash_site:}{Vector of crash site locations}
}
\value{
None: Prints average health score per site
}
\description{
: Computes and prints average health score for each crash site
}
\examples{
# Calculate average health by crash site
average_health_by_site(health_score, crash_site)
}
