% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{average_health_score_deceased}
\alias{average_health_score_deceased}
\title{: Calculate Average Health Score of Deceased}
\usage{
average_health_score_deceased(health_score, survival)
}
\arguments{
\item{health_score:}{Vector of health scores}

\item{survival:}{Vector of survival status}
}
\value{
None: Prints average health score of deceased
}
\description{
: Computes and prints the average health score of deceased passengers
}
\examples{
# Calculate average health score of deceased
average_health_score_deceased(health_score, survival)
}
