% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{calculate_survival_rate}
\alias{calculate_survival_rate}
\title{: Calculate Total Survival Rate}
\usage{
calculate_survival_rate(survival)
}
\arguments{
\item{survival:}{Vector of survival status (0=deceased, 1=survived)}
}
\value{
None: Prints total survivors and survival rate
}
\description{
: Computes and prints the total number of survivors and survival percentage
}
\examples{
# Calculate survival statistics
calculate_survival_rate(survival)
}
