% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{gender_wise_survival_ratio}
\alias{gender_wise_survival_ratio}
\title{: Calculate Gender-Wise Survival Ratio}
\usage{
gender_wise_survival_ratio(survival, gender)
}
\arguments{
\item{survival:}{Vector of survival status}

\item{gender:}{Vector of passenger genders}
}
\value{
None: Prints male and female survival rates
}
\description{
: Computes and prints survival percentages for male and female passengers
}
\examples{
# Calculate gender-wise survival ratios
gender_wise_survival_ratio(survival, gender)
}
