% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{health_category}
\alias{health_category}
\title{: Categorize Health Scores}
\usage{
health_category(health_score)
}
\arguments{
\item{health_score:}{Vector of health scores}
}
\value{
categories: Vector of health categories
}
\description{
: Categorizes health scores into Low, Medium, High and prints counts
}
\examples{
# Categorize health scores
health_cats <- health_category(health_score)
}
