% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{heaviest_survivor}
\alias{heaviest_survivor}
\title{: Find Heaviest Survivor}
\usage{
heaviest_survivor(weight, survival)
}
\arguments{
\item{weight:}{Vector of passenger weights}

\item{survival:}{Vector of survival status}
}
\value{
None: Prints weight of heaviest survivor
}
\description{
: Identifies and prints the weight of the heaviest surviving passenger
}
\examples{
# Find heaviest survivor
heaviest_survivor(weight, survival)
}
