% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{most_common_class_survived}
\alias{most_common_class_survived}
\title{: Most Common Class Among Survivors}
\usage{
most_common_class_survived(class, survival)
}
\arguments{
\item{class:}{Vector of passenger classes}

\item{survival:}{Vector of survival status}
}
\value{
None: Prints most common class
}
\description{
: Identifies and prints the most common passenger class among survivors
}
\examples{
# Find most common class among survivors
most_common_class_survived(class, survival)
}
