% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{print_full_summary}
\alias{print_full_summary}
\title{: Print Full Summary Report}
\usage{
print_full_summary(age, survival, gender, class, health_score, crash_site)
}
\arguments{
\item{age:}{Vector of passenger ages}

\item{survival:}{Vector of survival status}

\item{gender:}{Vector of passenger genders}

\item{class:}{Vector of passenger classes}

\item{health_score:}{Vector of health scores}

\item{crash_site:}{Vector of crash site locations}
}
\value{
None: Prints detailed summary
}
\description{
: Prints a comprehensive summary of crash analysis
}
\examples{
# Print full summary
print_full_summary(age, survival, gender, class, health_score, crash_site)
}
