% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{site_with_highest_survival}
\alias{site_with_highest_survival}
\title{: Site with Highest Survival}
\usage{
site_with_highest_survival(crash_site, survival)
}
\arguments{
\item{crash_site:}{Vector of crash site locations}

\item{survival:}{Vector of survival status}
}
\value{
None: Prints site with highest survivors
}
\description{
: Identifies and prints the crash site with the most survivors
}
\examples{
# Find site with highest survival
site_with_highest_survival(crash_site, survival)
}
