% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{survivors_by_crash_site}
\alias{survivors_by_crash_site}
\title{: Survivors by Crash Site}
\usage{
survivors_by_crash_site(crash_site, survival)
}
\arguments{
\item{crash_site:}{Vector of crash site locations}

\item{survival:}{Vector of survival status}
}
\value{
None: Prints survivor counts per crash site
}
\description{
: Counts and prints number of survivors at each crash site
}
\examples{
# Calculate survivors by crash site
survivors_by_crash_site(crash_site, survival)
}
