% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment_02_SparshKarna_23BDS1172.R
\name{youngest_first_class_survivor}
\alias{youngest_first_class_survivor}
\title{: Find Youngest First-Class Survivor}
\usage{
youngest_first_class_survivor(age, survival, class)
}
\arguments{
\item{age:}{Vector of passenger ages}

\item{survival:}{Vector of survival status}

\item{class:}{Vector of passenger classes}
}
\value{
None: Prints age of youngest first-class survivor
}
\description{
: Identifies and prints the age of the youngest first-class survivor
}
\examples{
# Find youngest first-class survivor
youngest_first_class_survivor(age, survival, class)
}
